Add endpoint for adding sources

+++ src/main/java/com/example/slovaro_backend/entity/Source.java
@@ -9,29 +9,29 @@ public class Source {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    int id;
+    Long id;
 
 
     //@ManyToOne()
     //@JoinColumn(name="userid")
     @Column(name="userid")
-    int userId;
+    Long userId;
     @Column(name = "name")
     String name;
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public int getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(int userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
